Created Auth service and edited Login page
Created Auth service which function is to communicate with the auth API to either get the token or communicate the error to the app.
If the login is succesfull, the service stores the returned token in localstorage.
If it isn't succesfull, the service returns an error.
In Login component, created a placeholder user which is used to store the email and password of the user, and two properties to control if the login has failed due to an error in the email or the user of if either the email or password are empty.