Added check login guard
Created a guard that checks if the user is logged in to the app to decide whether the user can access to home page or not.
If user isn't logged in, the guard redirects it to login page. If he is logged, then the guard let's user see the homepage.
Modified login component to navigate to home as expected due to routes changes.
Modified auth service since there was methods that were not necessary