Added validation to Login Form
Added directives to login form's inputs to validate wether both the email and password inputs are empty or not. If they are empty, a message is shown to let the user know about this error.
Email input has also a email format validator to prevent user from trying to login with an invalid email format.
Also, if the form isn't valid, the submit button will be disabled until all the inputs has a valid text in them.
Code refactoring has been made on the login component's logic since logic validation it's not necessary anymore.