Created mainpage component and edited homepage
Created mainPage component wich will be the main page of the app once the user is logged. It will display the current team and it's stats.
Modified home page to be the parent component of the app once the user is logged in.
Created home's child routes on routing module to allow routing-outlet inside homepage to show the respective components.
Modified navbar to allow logout from the app trough "Logout" button.

TODO: make navbar prettier with icons.