Created heroCard component
Created Hero card component wich will be the one who shows all the heroes in the team in the main page.
It displays the photo and hero name of each hero, and his stats. It also has a collapsible button wich deploys the personal data of the hero, and a button to drop him off the team.
Added bootstrap's js since it's necessary to collapse to work.
Refactored some code on hero api so it work as expected. CORS is still necessary because of backend's structure.
Added the heroCard tag to mainpage's html to show the component.
Note: herocard it's a dumb component, all the data is given from his parent. So it's an reutilizable component.