Added teamcard component
Created teamcard component wich function is showing the overall team stats, being the first stat the highest stat. It also shows the average weight and height of the team.
Created team model to validate if a team is truly a team.
Created a method in hero service to calculate the stats of the team whenever a new hero is added to the team.
Added imports in homepage to communicate both component since teamcard is a dumb component.