Created addhero page and addheroform component
Created the page to add heroes to the team.
Created a addheroform wich contains a simple form with an input to search for a hero name.
Added event handler into addhero page to search for the desired hero and store it in an array.
Modified navbar's links for it to work properly.
Modified home's child routes to add the addhero page to the routes.