Changes to herocard to make it a more reutilizable component
Created a new property to indicate the card type. This property is used in the template to render some divs depending on it's value, so the card is usable in both addhero page and mainpage.
Created output events to let the component communicate with it's parent whenever an user wants to either delete or add a hero.