Added herocard to addhero page
Added herocard import to addhero page's template.
Added a card to let user know that the hero was succesfully added to the team.
Added a method to addhero page to add a hero to the team when is selected.
Minor styles changes to mainpage and added the correspondant card type to the cards.
Minor change to button in addheroform since it's inner text wasn't correct.