Added warning card to addhero page
Added a warning card on add hero page wich is shown if there are too many heroes or villains in the team.
Modified heroservice to return the strings as expected, since there were conditions that weren't correct.
Added a condition to check if the heroe's alignment is '-', counting him as a villain.