Changes to hero service
Refactored calcstat method to work properly, since the averages weren't properly calculated.
Added a 'mode' parameter to calcstat() to indicate what's the function mode.
Caclstat now adapts to wether we are removing or adding a member to the team.
Edited deletehero function to update internal counters and to call calcstat to recalculate the stats of the team.